+++ package.json
@@ -1,6 +1,5 @@
 {
   "name": "@example/basics",
-  "type": "module",
   "version": "0.0.1",
   "private": true,
   "scripts": {
@@ -17,11 +16,13 @@
     "@astrojs/sitemap": "^1.0.0",
     "@astrojs/tailwind": "^2.1.2",
     "@astrojs/vercel": "^2.3.6",
-    "@types/node": "^18.7.6",
+    "@example/basics": "link:",
+    "@types/node": "^18.11.17",
     "astro": "^1.6.8",
     "mdast-util-to-string": "^3.1.0",
     "motion": "^10.14.3",
     "reading-time": "^1.5.0",
-    "tailwindcss": "^3.0.24"
+    "tailwindcss": "^3.0.24",
+    "ts-node": "^10.9.1"
   }
 }
