+++ src/components/SpotifyNowPlaying.astro
@@ -1,5 +1,5 @@
 ---
-import type { NowPlayingTrackResponse } from "../lib/spotify";
+import type { NowPlayingTrackResponse } from "../lib/now";
 
 const URL = import.meta.env.URL;
