+++ src/lib/spotify.ts
@@ -1,24 +1,9 @@
-export type track = {
-    name: string,
-    artists: { name: string }[],
-    album: { images: { url: string }[] },
-    external_urls: { spotify: string }
-};
-
-export type artist = {
+type Artist = {
     name: string,
     images: { url: string }[],
     external_urls: { spotify: string },
     followers: { total: number },
 }
-
-export type NowPlayingTrackResponse = {
-    isPlaying: boolean;
-    title: string;
-    artist: string;
-    url: string;
-    img: string;
-};
 
 export const getAccessToken = async () => {
     let refresh_token = import.meta.env.SPOTIFY_REFRESH_TOKEN;
@@ -42,7 +27,7 @@ export const getAccessToken = async () => {
     return response.json();
 }
 
-export const nowPlaying = async () => {
+const nowPlaying = async () => {
     const { access_token } = await getAccessToken();
 
     return fetch("https://api.spotify.com/v1/me/player/currently-playing", {
@@ -52,33 +37,5 @@ export const nowPlaying = async () => {
     });
 };
 
-export const get = async () => {
-    const response = await nowPlaying();
-    console.log(response)
-
-    if (response.status === 204 || response.status > 400) {
-        return new Response(JSON.stringify({ isPlaying: false }), {
-            status: 200,
-        });
-    }
-
-    const { item } = await response.json();
-
-    const track = {
-        isPlaying: true,
-        title: item.name,
-        artist: item.artists
-            .map((_artist: artist) => _artist.name)
-            .join(", "),
-        url: item.external_urls.spotify,
-        img: item.album.images[0].url,
-    };
-
-    return new Response(JSON.stringify(track), {
-        status: 200,
-    });
-};
-
-get();
-
-export { }; 
+export { nowPlaying };
+export type { Artist };
